server/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
appnope==0.1.3
astroid==2.15.6
asttokens==2.2.1
attrs==23.1.0
autobahn==19.11.2
autoflake==2.2.0
Automat==22.10.0
backcall==0.2.0
bcrypt==4.0.1
black==23.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
comm==0.1.4
constantly==15.1.0
cryptography==41.0.3
debugpy==1.6.7.post1
decorator==5.1.1
dill==0.3.7
dnspython==2.4.2
ecdsa==0.18.0
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.101.0
greenlet==2.0.2
h11==0.14.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
incremental==22.10.0
ipykernel==6.25.1
ipython==8.14.0
isort==5.12.0
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
kiteconnect==4.2.0
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
nest-asyncio==1.5.7
packaging==23.1
parso==0.8.3
passlib==1.7.4
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pyflakes==3.1.0
Pygments==2.16.1
pylint==2.17.5
pymongo==4.4.1
PyMySQL==1.1.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
pyzmq==25.1.1
requests==2.31.0
rsa==4.9
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
stack-data==0.6.2
starlette==0.27.0
tomli==2.0.1
tomlkit==0.12.1
tornado==6.3.3
traitlets==5.9.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
wcwidth==0.2.6
wrapt==1.15.0
zipp==3.16.2
zope.interface==6.0
